Back-End/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
joblib==1.4.2
openai==1.51.0
rank-bm25==0.2.2
reportlab==4.2.2
